Separate relation from other binary analysis
This is because this operation is not reflective. Furthermore, this
allows us to change relational operations as such:

    if (a < b < c) {
        // code here
    }

The smallest should always be on the left in chained comparisons,
so we may not have:

    if (a > b > c) {
        // code here
    }

This is to prevent confusion. The type checker should ensure this.
As such, the desugarer will only only check if the right side of the
comparison is a comparison itself.