Add debug printing of parse trace
I realise now that skipping of indentation should not be done in the
parse block. Instead, we somehow need to check the indentation at the
start of every parse function. Depending on the context, the indentation
should be either 0 (an inlined expression), or 0 or greater if they
are a line.