Re-add to token (`->`)
Need to re-evaluate the syntax of functions. Do we still want to assign
to a function or do we switch to using the to (`->`) token?

Also change the grammar of anonymous functions. The arguments are now
represented using a normal tuple. The type checker should check that
these are indeed identifiers.