In list of consecutive newlines, if dedent, place after first newline
Before, dedents were placed after all newlines. This however is
troublesome if we have multiple newlines followed by dedent(s). Namely,
it made it look like an statement or expression at the end of a block
was not newline terminated, by followed by a newline and then dedent
terminated.

Now, when the lexer encounteres multiple newlines, it only stores the
first one, and uses a counter to keep track of any other newlines. These
are then added to the list once another expression is encountered.