Create Type enum
This will not only come handy in the type checker but also during the
desugar stage.
For instance, when we assign to a value, we need to know whether it is
a tuple or not.
Instead of creating another custom alias in the desugarer, it is
probably better to re-use the types from the type checker (which is
still to be written).