Insert dedent(s) after first newline when dedentation detected
This is similar to the previous behaviour. If we detect multiple
newlines followed by a dedent, we effectively always insert the dedent
after the first newline.

This is achieved by caching newlines until a token is encountered.
Only when we encounter a new token do we check the current indentation
level and determine whether we should dedent or indent before inserting
more newlines.