[ci skip] Reorganise command line input and create separate CLI struct
I don't know why, but or some reasons it does not want to generate
output at the moment if no output argument is given to the command line,
even though it should default to making a Python file alongside the
Mamba file.
This behaviour is reflected in the pipeline tests, which do still pass.
When using the command line however it doesn't work, and I'm not sure
why but when I call println! nothing is shown in my console when the
command line tests are run.