Make statements in file and arguments of function type explicit
An function type now takes an array of arguments instead of a single
argument, which may be a tuple of types (representing an array of
arguments).
So now, if you want to have an function type which represents a function
which takes a tuple, you'd write:
((a,b)) -> c

Furthermore, statements are now parsed separately in a file instead of
being grouped in a file.