Use Hash of TypeName when hashing Type
- Use ActualTypeName in locations where we would never expect a union

This allows us to still have HashSet's of Type's, even though a Type
contains Functions which may return a HashSet of TypeNames.
Normally, we cannot hash HashSet's, but this allows us to circumvent
this.

Furthermore, this also makes more sense from a conceptual level.
Types that have the same name are the same by definition.
If we have multiple definitions of the same type, we should raise an
error whilst building the context.