Use Sets where ordering is not relevant conceptually
We still need to make our type system interface with Python's type hint
system. For instance, the for loop test is now failing as the built-in
set type expects no generic arguments, even though it should accept one
generic.

As such, we need to read the typing library of Python before proceeding:
https://docs.python.org/3/library/typing.html