Add desugar rule to end function with explicit return type with return
It is a slightly less elegant solution, but until we have implemented a
system which allows the type checker to augment the ast this is the best
solution at hand.
Otherwise, adding more complex desugar rules to work in parallel with
the type checker can quickly become unmanageable for more advanced
language features.

It should be noted that the expect an expression boolean is passed on
recursively, which can result in ternary operators appearing in
unexpected locations in malformed applications (which highlights the
point I made above).