Use modified output from type checker in pipeline
If constructor call, modify state during desugar stage.
Perhaps we can modify the AST within the type checker, so it doesn't
awkwardly rely on a function within the type checker.

While writing this I also noted how sphagetti like the application logic
has become for call constructors.
I identified a lot of application logic which could be streamlined now
that we have a modification pass in the type checker, so we don't need
to do a lot of work in the desugar stage.

We also now have the issue that something is treated as a constructor
even if there is a local function with the same name.
We either need to change this, or if we want to keep this (intended
behaviour) produce an appropriate warning/error message.