Create simple constraint struct
We still need to have a system for constraints between two AST.
Namely, we need to check that the type of an expression is equal to that
of another expression.
So we need a constraint between two separate Expressions/ASTs.

In future we also likely need to add constraints for functions.