Implement Implements constraint
This is done by checking whether a type defines a function, and then
proceeding to binding these function arguments and then unifying these
arguments.
It may be possible to move this logic to the generation stage as we are
generating new constraints within the unification stage which is a bit
confusing.

- Add equality to String class
- Remove mutable boolean from environment. We wrap a mutable expect in
  a mutable