Use recursion in unify
This prevents us from accidentally creating new lists
and ignoring the old ones.
Note that in some situations we grow the constraint
list, but whenever we do this we take care to append
it to existing constraints.

Use a while loop in the substitution phase, since that
does make more sense there then to do it recursively.