Constrain based on assign or return type
We now set the expected Expression Type
in the state such that we can constrain
inner expressions based on the expected
final type.

We will have to reset the state in some
instances, since it is now done recursively
for all expressions, i.e. all expressions
in a block are constrained in this manner
instead of only the last one.