Use VecDeque in Constraints
New constraints are pushed at the back and new
ones popped from the front.

Originally, the intention was that we can reinsert
constraints into the queue if we cannot sole them at
that time.
One such constraint is HasField and Expression.
The idea is that we reinsert this in the queue, and
hopefully the Expression has been substituted with
a Type by the time we reincounter the constraint.

I think we need to add a boolean flag to each
constraint to indicate that it has been reinserted,
so that we can detect infinite loops.
(Kind of like how we can solve the Halting problem if
a machine is linearly bound, kind of, or that's where
I got the idea from)