Set up framework for sets of constraints
Each set is its own universe of constraints.
The codebase has lost some of its elegance however,
as we are currently modifying a mutable reference
in addition to returning a copy.

I also need to take care of what I do with the
environment in some situations.
As it stands now a lot of variables are global
scope when they shouldn't be, and I don't do
a unification of the environment where relevant,
e.g. after an if statement.