Specify left side of constraint as parent
This means that the type of the left hand side
must be a superset of the right hand side.
See is_superset method in TypeName, in short:
- left hand side may be nullable even if right
  hand side isn't, but not other way around.
- left hand side must be a typing union that is
  greater than right hand side.

Should look into ways to clearly document this and
perhaps modify the constructor since this may
not be so clear during the constraint generation
stage.