Fix constraint gen for nested reassign
- Don't clone substitute, there was no reason to
  do this and it caused a lot of duplicate
  substitutions to take place

We should consider how we handle top-level variable
definitions in classes.
I think that at the very least if they are public,
then their type should be explicit.
However, this also means that we have to assign a type
even if it is trivially evident what the type is.

I would propose that in future, the context be updated
so it generates a set of constraints for each class.
However, the constraints generated are only those
relevant to the public interface of the class, so we
should not reveal any internal logic.