Check raises within functions
- All raised expressions must have one
  of the types defined in the function
  signature
- Top-level raises within a script are not
  constrained
- All type names in the function definition
  which are listed as possible raises are now
  verified to inherit from Exception

The same application logic used

Also uncovered a bug where classes with only
class arguments and nothing else result in an
empty constructor.
This is indicative of the convoluted nature of
the desugar stage at this point, which is not
logically set up.
Much of this application logic should be moved
to the modification stage in the type checker
so we can progressively build classes instead of
trying to do everything in one go which is fickle.