Add extra check to detect circular parents
This should be checked immediately after
the Context is built however.

There was an issue where it didn't terminate.
Now it does.
This didn't have anything to do with circular
dependencies, it had something to do with the
for loops within unify_link.rs, though I am not
sure what it was. Nevertheless, the issue has
been rectified.