Search direct fields and functions first
We always check the type of the current
field and function first before going onto
the fields and functions of parents.

We should formalise this in future in the
language specification.
I think that it is best that when building
the context the function signatures of
functions overwritten by children are identical
to the signature of the parent.