Use identifier flags for scoping
We now add identifiers to constraints to
flag up until what point we should
substituted.

However, I have not checked how this
influences re-insert constraints.
I think we should rework this system such
that we always preserve order, so we just
have to do a second pass.

This will break programs where we only
re-define parts of tuples, if the redefined
parts have a different type.