Start simplfying Type and TypeName
We are moving towards an architecture where
Type does not contain an entire copy of a
Class within, but rather a copy of a Name,
which we can use to look up classes within
the Context.

Not only does this save a lot of memory, but
conceptually this makes more sense.