Use environment to detect shadowed variables
We substitute variables which shadow a previous
variable with a variable which the user cannot
create (because it contains an illegal character).

We also store this substitution such that any
other uses of this variable which shadowed the
prior one get the new name.

This is far more elegant than the ad-hoc system
of storing variables, and removes some of the
complexity away from the constraints.