[ci skip] Extra tuple substitution rules
In the unification stage that is.
Not sure if these are strigly necessary,
or if this is a hack which will resolve
itself once we hopefully land on an
elegant solution.

It does appear to fix some substitution
issues.
However, we again run into the issues that
trivial constraints (such as primitive
expressions) are no longer generated in the
unification stage.

Ideally, all constraints should be
generated in the generation stage, and
the unification stage should solve these
as-is.