Fix unification of tuples
The fix was: first substitute,
then proceed to extract the
tuple elements as necessary.
I think this is the only type of
expression (for now) where we need
to extract the individual elements.
All other expressions can be
dealt with as-is, and should be
simple enough that all constraints
are generated during the generation
stage.

We also ignore two tuples,
and just let the substitution do
its work in this case.

- Update some error msg formatting
- Simplify unify expression rules