Add var mapping logic to union and intersect
Also move environment to generation stage, as it
is only used here.
It does not need to be exposed to the outside
world.

Expect(ed) now takes a mapping of variables, which
is a mapping of a string to a string, as it is all
that it needs for shadowing purposes.
We do not need to expose the entire environment to
Expect(ed).