[ci skip] none is seen as a superset of nullable
There appear to be issues with how constraints are
implemented.
The left side should always be the superset of the
right side of a constraint in a correctly typed
input.
However, we get type errors for correctly typed
applications, implying that constraints are not
properly being generated in these cases.