Access may be either left or right
This appears to have been the root cause of the
problem.
Now, when the access is either on the left
or right side of the constraint, unification is
done using the proper side.
Before, we got an overflow or an error because
we always assumed the access was on the left
side, even if it was not the case.