Streamline some type checker tests
We now return results here, as was suggested in
#63.
The results are also pretty printed.

- Move some unit tests to proper locations,
  alongside the source.
- Always delete temporary directory before
  continuing with output tests.
  This should fix the issue of constant temporary
  directories when tests fail.

We need to move the unit tests to the appropriate
locations for the remainder of the test suite.
Same for the more elegant tests and pretty print,
this will really help with development in the
future.