Start adding logic for calls
This would fix a lot of erroneus behaviour we have
encountered.
We even have an issue for one:
- #188

(Lets ignore the fact that I forgot to make a few
issues as I went along for some other weird
behaviour).

The current implementation therefore is quite
brittle.
The type checker currently flat out ignores calls
in certain situations (such as reassigns it seems)
and just assumes we are assigning to the variable
at the front of the chain directly, while we are
actually asigning to a (deeply) nested field.

If we manage to fix this, then:
- Nested calls should be properly type checked, no
  matter how deep they are
- In nested calls, we don't check the environment
  but instead the class (via the context).
  We can do this by either:
  - Creating an access constraint on the fly as we
    go deeper
  - Looking up the field in the class directly.
    This seems more efficient, and because all
    signature do specify the class this should be
    straightforward without the need for any
    unification.