Add optional to Name in generate
Always print out generics in alphabetical order.
This ensures that the output is easy to predict,
we are not dependent on the whims of the iterator
over the set, possibly preventing flaky tests.

Make annotating output the default in tests as
well. As we start to implement new features and
debug the checking stage we should then
automatically see old tests failing as well.