Fix redundant ABC
If a class has a parent which is already abstract,
then we don't need to mark this class as abstract.
We do this check recursively.

Also added context to generate stage to perform
this check.
More advanced language constructs could benefit
from a context in the generate stage in future.