Automatically convert last expression to return
As mentioned in #164.

However, also exposed a weird error in check
stage. In `function/definition.mamba`, line 7:
Change
`    if True then return 10 else return None`
To
`    if True then 10 else None`
Gives type error:
`Unifying two types: Expected an Int, was a None`
Not something to fix in this PR, it is erroneous
behaviour as the mamba script is well-typed
according to current spec.