[skip] verify raises is parsed correctly
This confirms that if else at end of function
def (definition_ast_verify) is likely caused by
unification with function definition.
Ideally we want to leave the generate stage as-is.
The mapping within the checker should be smart
enough to see that the if contains branches with
explicit returns.
If this is inconsistent (one branch contains
return, other does not), checker should give an
error.