Ternary expressions (#364)
* Generate new constr if if_expr used as expression

- Streamline ExpressionAny in check stage
- Add tests for if constraint generations

In generate stage we will need to for all
constructs check whether something expects an
expression. This is given by the exp_expression
flag.

* Don't assign type in with resource construct

Types don't properly propagate out of if
expressions.

If's are also being assigned types
when they are in fact statements in most places it
appears. This is likely because we store something
in the environment which should in fact not
propagate inwards.

* Don't propagate exp_expression out of fun arg

Same for identifier from var.

* Add constraint for undefined

Properly treat it as a type.

* [skip] verify raises is parsed correctly

This confirms that if else at end of function
def (definition_ast_verify) is likely caused by
unification with function definition.
Ideally we want to leave the generate stage as-is.
The mapping within the checker should be smart
enough to see that the if contains branches with
explicit returns.
If this is inconsistent (one branch contains
return, other does not), checker should give an
error.

* Propagate assign to if and match

We will not fix the introduced bug here, see #366.
Will fix in separate PR which also tests other
similar situations where a function should not
check in a more exhaustive manner.

* Fix constraint position for for loop collection

- Add Expected::any
- Ignore successfully unified types if they are
  `Any`.
- Uncover some issues highlighted in #37

* Use new let else from Rust 1.65.0

- Add binary one compliment test as well for cov
- Streamline some generate logic

* Remove some redundant logic

- Reset some logic for with resource gen