[ci skip] Fix field access constraint generation
However now we have an issue that primitive
comparisons seems to trip up.

Something is up with how we generate constraints
on the fly for field access in conjunction with
function access.

If we flip back the field_ty_exp, then we no
longer get the failed definition_ast_verify test.
It seems that if we fix the issue with the Union,
we get nullability check issues and vice versa.