Add exception that Name may be super of any
This means that in certain situations, a Name is
a superset of another if it is a superset of just
one of the types of other.

For now, this is useful at a function call site
where we generate such constraints on the fly.
There, we only need the current expression to be
super of one.

However, seeing as we only have one use case at
the moment, I'm not sure this logic belongs in
Name itself.
Perhaps this might be useful in future for fixing
other issues.
Issues such as #296 come to mind.

- Unignore certain mutability test.