[ci skip] Index now treated as StringName
In certain contexts.
Also substitute generics in function calls, which
includes constructors.

Only issue now is that we must mark generics when
checking class bodies.
This means that we should check whether they exist
in the context, but instead whether they are
mentioned as a generic in the class name.

The two new tests hint as this, but within this PR
the generate stage should also preserve the
generics of classes in the output.